+++ bar/bar/database.c
@@ -1514,7 +1514,7 @@ void Database_doneAll(void)
                       )
 #else /* not NDEBUG */
   Errors __Database_open(const char        *__fileName__,
-                         uint              __lineNb__,
+                         ulong             __lineNb__,
                          DatabaseHandle    *databaseHandle,
                          const char        *fileName,
                          DatabaseOpenModes databaseOpenMode,
@@ -1712,7 +1712,7 @@ void Database_doneAll(void)
   void Database_close(DatabaseHandle *databaseHandle)
 #else /* not NDEBUG */
   void __Database_close(const char   *__fileName__,
-                        uint         __lineNb__,
+                        ulong        __lineNb__,
                         DatabaseHandle *databaseHandle
                        )
 #endif /* NDEBUG */
@@ -1933,7 +1933,7 @@ dumpRequest("yield",databaseHandle);
   void Database_lock(DatabaseHandle *databaseHandle)
 #else /* not NDEBUG */
   void __Database_lock(const char   *__fileName__,
-                       uint         __lineNb__,
+                       ulong        __lineNb__,
                        DatabaseHandle *databaseHandle
                       )
 #endif /* NDEBUG */
@@ -1957,7 +1957,7 @@ dumpRequest("yield",databaseHandle);
   void Database_unlock(DatabaseHandle *databaseHandle)
 #else /* not NDEBUG */
   void __Database_unlock(const char   *__fileName__,
-                         uint         __lineNb__,
+                         ulong        __lineNb__,
                          DatabaseHandle *databaseHandle
                         )
 #endif /* NDEBUG */
@@ -3158,7 +3158,7 @@ Errors Database_removeColumn(DatabaseHandle *databaseHandle,
   Errors Database_beginTransaction(DatabaseHandle *databaseHandle)
 #else /* not NDEBUG */
   Errors __Database_beginTransaction(const char   *__fileName__,
-                                     uint         __lineNb__,
+                                     ulong        __lineNb__,
                                      DatabaseHandle *databaseHandle
                                     )
 #endif /* NDEBUG */
@@ -3387,7 +3387,7 @@ Errors Database_execute(DatabaseHandle      *databaseHandle,
                          )
 #else /* not NDEBUG */
   Errors __Database_prepare(const char          *__fileName__,
-                            uint                __lineNb__,
+                            ulong               __lineNb__,
                             DatabaseQueryHandle *databaseQueryHandle,
                             DatabaseHandle      *databaseHandle,
                             const char          *command,
@@ -3748,7 +3748,7 @@ bool Database_getNextRow(DatabaseQueryHandle *databaseQueryHandle,
   void Database_finalize(DatabaseQueryHandle *databaseQueryHandle)
 #else /* not NDEBUG */
   void __Database_finalize(const char        *__fileName__,
-                           uint              __lineNb__,
+                           ulong             __lineNb__,
                            DatabaseQueryHandle *databaseQueryHandle
                           )
 #endif /* NDEBUG */
