+++ bar/bar/deltasourcelists.h
@@ -109,7 +109,7 @@ void DeltaSourceList_doneAll(void);
   void DeltaSourceList_init(DeltaSourceList *deltaSourceList);
 #else /* not NDEBUG */
   void __DeltaSourceList_init(const char      *__fileName__,
-                              uint            __lineNb__,
+                              ulong           __lineNb__,
                               DeltaSourceList *deltaSourceList
                              );
 #endif /* NDEBUG */
@@ -133,8 +133,8 @@ void DeltaSourceList_doneAll(void);
                                      const DeltaSourceNode *fromDeltaSourceListToNode
                                     );
 #else /* not NDEBUG */
-  void __DeltaSourceList_initDuplicate(const char        *__fileName__,
-                                       uint              __lineNb__,
+  void __DeltaSourceList_initDuplicate(const char            *__fileName__,
+                                       ulong                 __lineNb__,
                                        DeltaSourceList       *deltaSourceList,
                                        const DeltaSourceList *fromDeltaSourceList,
                                        const DeltaSourceNode *fromDeltaSourceListFromNode,
@@ -155,7 +155,7 @@ void DeltaSourceList_doneAll(void);
   void DeltaSourceList_done(DeltaSourceList *deltaSourceList);
 #else /* not NDEBUG */
   void __DeltaSourceList_done(const char      *__fileName__,
-                              uint            __lineNb__,
+                              ulong           __lineNb__,
                               DeltaSourceList *deltaSourceList
                              );
 #endif /* NDEBUG */
