+++ bar/bar/index.c
@@ -216,7 +216,7 @@ LOCAL bool                       quitFlag;
                         )
 #else /* not NDEBUG */
   LOCAL Errors __openIndex(const char  *__fileName__,
-                           uint        __lineNb__,
+                           ulong       __lineNb__,
                            IndexHandle *indexHandle,
                            const char  *databaseFileName,
                            uint        indexOpenModes,
@@ -316,7 +316,7 @@ LOCAL bool                       quitFlag;
   LOCAL Errors closeIndex(IndexHandle *indexHandle)
 #else /* not NDEBUG */
   LOCAL Errors __closeIndex(const char  *__fileName__,
-                            uint        __lineNb__,
+                            ulong       __lineNb__,
                             IndexHandle *indexHandle
                            )
 #endif /* NDEBUG */
@@ -4198,7 +4198,7 @@ IndexHandle *Index_open(uint priority,
                        )
 #else /* not NDEBUG */
 IndexHandle *__Index_open(const char *__fileName__,
-                          uint       __lineNb__,
+                          ulong      __lineNb__,
                           uint       priority,
                           long       timeout
                          )
@@ -4272,7 +4272,7 @@ void Index_interrupt(IndexHandle *indexHandle)
 Errors Index_beginTransaction(IndexHandle *indexHandle, ulong timeout)
 #else /* not NDEBUG */
 Errors __Index_beginTransaction(const char  *__fileName__,
-                                uint        __lineNb__,
+                                ulong       __lineNb__,
                                 IndexHandle *indexHandle,
                                 ulong       timeout
                                )
